- Move Python variables from Job to Bacula. They are   DirName, Version, ConfigFile, and WorkingDir - Fix delete of bootstrap to only occur on Bacula created   filenames. - Allow cancelling a Job name that is not active. It is   sent to the daemons.
git-svn-id: https://bacula.svn.sourceforge.net/svnroot/bacula/trunk@2210 91ce42f0-d328-0410-95d8-f526ca767f89

+++ bacula/src/dird/fd_cmds.c
@@ -449,7 +449,10 @@ bool send_bootstrap_file(JCR *jcr)
    }
    bnet_sig(fd, BNET_EOD);
    fclose(bs);
-   unlink(jcr->RestoreBootstrap);
+   if (jcr->unlink_bsr) {
+      unlink(jcr->RestoreBootstrap);
+      jcr->unlink_bsr = false;
+   }                         
    if (!response(jcr, fd, OKbootstrap, "Bootstrap", DISPLAY_ERROR)) {
       set_jcr_job_status(jcr, JS_ErrorTerminated);
       return 0;
